Remove placeholder 'REVISION' text objects from schematics for ALU and register boards
These originally existed so that a simple text find-and-replace could be run over all schematics to generate files labelled e.g. "Revision 1". Since published revisions are not going to be done, it makes sense to remove these
placeholder labels. Only the ALU and register boards have had these removed, since the prog and gpio boards don't have labels yet, and there appears to have been a mixup when importing the ctrl and mem boards into the repo which
will be resolved later.